1.4.6

+++ ShinRyuModManager-CE/Program.cs
@@ -199,38 +199,40 @@ public static class Program {
         }
         
         Log.Information("Active Profile: {Profile}", profile?.GetDescription() ?? ActiveProfile.GetDescription());
-        
-        // TODO: Maybe move this to a separate "Game patches" file
-        // Virtua Fighter eSports crashes when used with dinput8.dll as the ASI loader
-        if (GamePath.CurrentGame == Game.Eve && File.Exists(Constants.DINPUT8DLL)) {
-            if (File.Exists(Constants.VERSIONDLL)) {
-                Log.Warning($"Game specific patch: Deleting {Constants.DINPUT8DLL} because {Constants.VERSIONDLL} exists...");
-                
-                // Remove dinput8.dll
-                File.Delete(Constants.DINPUT8DLL);
-            } else {
-                Log.Warning($"Game specific patch: Renaming {Constants.DINPUT8DLL} to {Constants.VERSIONDLL}...");
-                
-                // Rename dinput8.dll to version.dll to prevent the game from crashing
-                File.Move(Constants.DINPUT8DLL, Constants.VERSIONDLL);
-            }
-        } else if (GamePath.CurrentGame is >= Game.YakuzaKiwami2 and not Game.LikeADragonGaiden) {
-            // Lost Judgment (and Judgment post update 1) does not like Ultimate ASI Loader, so instead we use a custom build of DllSpoofer (https://github.com/Kazurin-775/DllSpoofer)
-            if (File.Exists(Constants.DINPUT8DLL)) {
-                Log.Warning($"Game specific patch: Deleting {Constants.DINPUT8DLL} because it causes crashes with Judgment games...");
-                
-                // Remove dinput8.dll
-                File.Delete(Constants.DINPUT8DLL);
-            }
-            
-            if (!File.Exists(Constants.WINMMDLL)) {
-                if (File.Exists(Constants.WINMMLJ)) {
-                    Log.Warning($"Game specific patch: Enabling {Constants.WINMMDLL} by renaming {Constants.WINMMLJ} to fix Judgment games crashes...");
+
+        if (File.Exists(Constants.DINPUT8DLL) && File.Exists(Constants.WINMMLJ)) {
+            // TODO: Maybe move this to a separate "Game patches" file
+            // Virtua Fighter eSports crashes when used with dinput8.dll as the ASI loader
+            if (GamePath.CurrentGame == Game.Eve && File.Exists(Constants.DINPUT8DLL)) {
+                if (File.Exists(Constants.VERSIONDLL)) {
+                    Log.Warning($"Game specific patch: Deleting {Constants.DINPUT8DLL} because {Constants.VERSIONDLL} exists...");
+                    
+                    // Remove dinput8.dll
+                    File.Delete(Constants.DINPUT8DLL);
+                } else {
+                    Log.Warning($"Game specific patch: Renaming {Constants.DINPUT8DLL} to {Constants.VERSIONDLL}...");
                     
                     // Rename dinput8.dll to version.dll to prevent the game from crashing
-                    File.Move(Constants.WINMMLJ, Constants.WINMMDLL);
-                } else {
-                    Log.Error($"WARNING: {Constants.WINMMLJ} was not found. Judgment games will NOT load mods without this file. Please redownload Shin Ryu Mod Manager.");
+                    File.Move(Constants.DINPUT8DLL, Constants.VERSIONDLL);
+                }
+            } else if (GamePath.CurrentGame is >= Game.YakuzaKiwami2 and not Game.LikeADragonGaiden) {
+                // Lost Judgment (and Judgment post update 1) does not like Ultimate ASI Loader, so instead we use a custom build of DllSpoofer (https://github.com/Kazurin-775/DllSpoofer)
+                if (File.Exists(Constants.DINPUT8DLL)) {
+                    Log.Warning($"Game specific patch: Deleting {Constants.DINPUT8DLL} because it causes crashes with Judgment games...");
+                    
+                    // Remove dinput8.dll
+                    File.Delete(Constants.DINPUT8DLL);
+                }
+                
+                if (!File.Exists(Constants.WINMMDLL)) {
+                    if (File.Exists(Constants.WINMMLJ)) {
+                        Log.Warning($"Game specific patch: Enabling {Constants.WINMMDLL} by renaming {Constants.WINMMLJ} to fix Judgment games crashes...");
+                        
+                        // Rename dinput8.dll to version.dll to prevent the game from crashing
+                        File.Move(Constants.WINMMLJ, Constants.WINMMDLL);
+                    } else {
+                        Log.Error($"WARNING: {Constants.WINMMLJ} was not found. Judgment games will NOT load mods without this file. Please redownload Shin Ryu Mod Manager.");
+                    }
                 }
             }
         }
